requirements.txt
aiosqlite==0.11.0
ciso8601==2.1.3
click==7.1.1
fastapi==0.53.2
fastapi-utils==0.2.1
h11==0.9.0
httptools==0.1.1
pydantic==1.4
PyPika==0.36.0
SQLAlchemy==1.3.15
starlette==0.13.2
tortoise-orm==0.16.3
typing-extensions==3.7.4.1
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
